HP Bar, Enemy Chaser, Damage Mechanism, Engineer Sprite
Added a functional "Slider" HP bar that decreases when player is damaged with it's respective script "HealthSlider" to take care of its functionality. Implemented a script named "EnemyChasePlayer" which can be used on any enemy to chase the player. "EnemyChasePlayer" depends heavily on two "Capsule Colliders (2D)" to function properly. One collider acts like a line of sight which triggers the enemy to give chase to the player. The other collider is to physically check when the enemy comes into contact with another collider (physical object).

The "EnemyDamage" script can be used to customize the amount of damage each "collision" does to the player. This is done by searching for a "HealthSlider" class on the object it collides with. If it collides with the player it will inflict the specified "damage" the "health" variable via the "TakeDamage" function located in the script/class "HealthSlider".

A larva monster was left on the first room entered after the hangar to test all these scripts. It is fully equipped with a "damaging" mechanism and can chase the player via the "EnemyChasePlayer" script. The larva can be duplicated as many times as needed while fully scripted and set up to save time while placing enemies.

Finally, an engineer sprite was added for future use.